+++ cmd/nodecmd/upgrade.go
@@ -5,6 +5,7 @@ package nodecmd
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/ava-labs/avalanche-cli/pkg/dependencies"
 	"strings"
 	"sync"
 
@@ -185,7 +186,7 @@ func getNodesUpgradeInfo(hosts []*models.Host) (map[*models.Host]nodeUpgradeInfo
 					nodeUpgradeInfo.SubnetEVMIDsToUpgrade = append(nodeUpgradeInfo.SubnetEVMIDsToUpgrade, vmName)
 				}
 				// find the highest version of avalanche go that is still compatible with current highest rpc
-				avalancheGoVersionToUpdateTo, err = node.GetLatestAvagoVersionForRPC(app, rpcVersion, latestAvagoVersion)
+				avalancheGoVersionToUpdateTo, err = dependencies.GetLatestAvalancheGoByProtocolVersion(app, rpcVersion)
 				if err != nil {
 					nodeErrors[hostID] = err
 					continue
